Move the authorization code flow into its own component
